+++ app/resources/views/layouts/auth.blade.php
@@ -41,6 +41,7 @@
     </head>
     <body class="antialiased">
         @yield('content')
+        @include('partials.site-footer')
         @stack('scripts')
     </body>
 </html>